requirements.txt
langchain
langchain-community
langchain-google-genai
psycopg2-binary
python-dotenv
streamlit